+++ src/main/java/chess/service/ChessGameService.java
@@ -1,6 +1,7 @@
 package chess.service;
 
 import chess.dao.ChessGameDao;
+import chess.dao.PieceDao;
 import chess.domain.ChessGame;
 import chess.domain.GameResult;
 import chess.domain.generator.BlackGenerator;
@@ -10,6 +11,7 @@ import chess.domain.player.Team;
 import chess.domain.position.Position;
 import chess.dto.ChessGameDto;
 import chess.dto.ChessGameUpdateDto;
+import chess.dto.PieceDto;
 import chess.dto.StatusDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
@@ -18,9 +20,11 @@ import org.springframework.stereotype.Service;
 public class ChessGameService {
 
     private final ChessGameDao chessGameDao;
+    private final PieceDao pieceDao;
 
-    public ChessGameService(ChessGameDao chessGameDao) {
+    public ChessGameService(ChessGameDao chessGameDao, PieceDao pieceDao) {
         this.chessGameDao = chessGameDao;
+        this.pieceDao = pieceDao;
     }
 
     public ChessGameDto createNewChessGame(final String gameName) {
@@ -28,11 +32,18 @@ public class ChessGameService {
         if (isDuplicate) {
             throw new IllegalArgumentException("중복된 게임 이름입니다.");
         }
+        final ChessGame chessGame = initializeChessGame();
+        final int chessGameId = chessGameDao.createNewChessGame(chessGame, gameName);
+        pieceDao.savePieces(chessGame.getCurrentPlayer(), chessGameId);
+        pieceDao.savePieces(chessGame.getOpponentPlayer(), chessGameId);
+        return ChessGameDto.of(chessGame, gameName);
+    }
+
+    private ChessGame initializeChessGame() {
         final Player whitePlayer = new Player(new WhiteGenerator(), Team.WHITE);
         final Player blackPlayer = new Player(new BlackGenerator(), Team.BLACK);
         final ChessGame chessGame = new ChessGame(whitePlayer, blackPlayer);
-        chessGameDao.createNewChessGame(chessGame, gameName);
-        return ChessGameDto.of(chessGame, gameName);
+        return chessGame;
     }
 
     public StatusDto findStatus(final String gameName) {
@@ -46,7 +57,7 @@ public class ChessGameService {
     public StatusDto finishGame(final String gameName) {
         final StatusDto status = findStatus(gameName);
         final int gameId = findGameIdByGameName(gameName);
-        chessGameDao.deletePieces(gameId);
+        pieceDao.deletePieces(gameId);
         chessGameDao.deleteChessGame(gameId);
         return status;
     }
@@ -63,15 +74,17 @@ public class ChessGameService {
         final Player opponentPlayer = chessGame.getOpponentPlayer();
         chessGame.move(currentPlayer, opponentPlayer, new Position(current), new Position(destination));
         chessGameDao.updateGameTurn(gameId, chessGame.getTurn());
-        chessGameDao.updatePiece(gameId, current, destination, currentPlayer.getTeamName(),
+        pieceDao.updatePiece(gameId, current, destination, currentPlayer.getTeamName(),
                 opponentPlayer.getTeamName());
         return ChessGameDto.of(chessGame, gameName);
     }
 
     private ChessGame findGameByName(final String gameName) {
         final int chessGameId = findGameIdByGameName(gameName);
-        final ChessGameUpdateDto gameUpdateDto = chessGameDao.findChessGame(chessGameId);
-        return ChessGame.of(gameUpdateDto);
+        final String currentTurn = chessGameDao.findCurrentTurn(chessGameId);
+        final List<PieceDto> whitePieces = pieceDao.findAllPieceByIdAndTeam(chessGameId, Team.WHITE.getName());
+        final List<PieceDto> blackPieces = pieceDao.findAllPieceByIdAndTeam(chessGameId, Team.BLACK.getName());
+        return ChessGame.of(new ChessGameUpdateDto(currentTurn, whitePieces, blackPieces));
     }
 
     private int findGameIdByGameName(final String gameName) {
